router: Move RejectionRecorder to its own file
Move the RejectionRecorder interface and default implementation from
unique_host.go into its own file, rejection_recorder.go.
